+++ packages/components/src/shared/types.ts
@@ -52,7 +52,7 @@ export interface InputDOMProps extends IdProps, InputDOMEvents {
     | "search"
 }
 
-interface DOMElement extends Element, HTMLOrSVGElement {}
+export interface DOMElement extends Element, HTMLOrSVGElement {}
 
 type DataAttributes = {
   [dataAttr: string]: any